us_market/requirements.txt
Flask
pandas
numpy
yfinance
requests
python-dotenv
google-generativeai
openai
ta
tqdm
